Switched long long int to int64_t, same for unsigned

+++ src/apps/dggrid/DgHexSF.h
@@ -91,11 +91,11 @@ class DgHexSF {
       bool operator!= (const DgHexSF& h) const
             { return !operator==(h); }
 
-      unsigned long long int depthFirstTraversal (GridGenParam& dp, 
+      uint64_t depthFirstTraversal (GridGenParam& dp, 
                   const DgIDGG& dgg, const DgContCartRF& deg, int numAp4Res,
                   DgEvalData* ed = NULL);
 
-      unsigned long long int visitMe (GridGenParam& dp, const DgIDGG& dgg,
+      uint64_t visitMe (GridGenParam& dp, const DgIDGG& dgg,
                           const DgContCartRF& deg, DgEvalData* ed);
 
       DgHexSF downAp4 (void);
